fix: 샘플링 메서드 관련 테스트 코드 수정 (#152)
